requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
annoy==1.17.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attr==0.3.2
attrs==22.1.0
blis==0.7.8
boto3==1.26.6
botocore==1.29.6
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
datasets==2.6.1
dill==0.3.5.1
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.39.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.3
gensim==4.2.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
hnswlib==0.6.2
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.12.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
line-profiler==4.0.2
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.9
numpy==1.23.4
oauthlib==3.2.1
omegaconf==2.3.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.8
protobuf==3.19.6
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.2
similarities==1.0.3
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text2vec==1.1.7
thinc==8.1.4
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.13.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.1
transformers==4.23.1
typer==0.4.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.1
zipp==3.9.0
